Update application-gateway-autoscaling-zone-redundant.md
we know that even configured minimum instance as 0, there are always 2 instance running internally. Thus there is no such difference configured minimum instance count from 0 to 2. So the downtime of scaling out “from 0 to 5” and “from 2 to 5” will be the same. I suppose that we should edit this part minimum instance count of at least 3 instead.